Refactor SwiftUI components
- Change animation from `spring` to `.easeInOut` when changing application states
- Remove `ContentView` and move `ContentView.Action`'s to `ContentListView'
- Remove redudant SwiftUI views and simplify composition by removing
  `VStack`'s and `HStack`'s
- Move code related to filtering into its own view called `ContentListFilterView`
- Remove being able to publish `nil` values in the `ViewModelsPublisher`
- Move adding groups view into `SidebarAddGroupButtonView`
- Move detail background into `SingleDetailBackgroundView`
